fix(routing): fix session_epoch type inconsistency and validation logic
- Unify session_epoch type to uint32_t across all structures
- Fix epoch validation condition: check session_id != 0 only (epoch 0 is valid)
- Update comments for clarity

This addresses the review findings from the session epoch validation
implementation, ensuring type consistency and correct validation logic.

+++ include/shield/net/session.hpp
@@ -52,7 +52,8 @@ struct ServiceAddress {
 struct SessionRoutingContext {
     std::string gateway_address;
     std::string session_id;
-    uint64_t session_epoch = 0;
+    uint32_t session_epoch =
+        0;  // Changed from uint64_t to uint32_t for consistency
     std::string player_id;
     std::string protocol_profile_id;
     std::unordered_map<std::string, ServiceAddress> service_routes;
